+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `ps` command shows a list of services and tasks currently active in a cluster
+
+
 ### Removed
 
 - No longer assign public IP address or allow public/private subnet config